requirements.txt
# --- Core Web Framework ---
fastapi
# The modern, high-performance web framework for building your API.

uvicorn[standard]
# The web server that runs your FastAPI application.
# We use [standard] to include performance-enhancing libraries.


# --- Database & ORM (Object-Relational Mapper) ---
sqlmodel
# The library that lets you interact with your database using Python classes.
# It cleverly combines SQLAlchemy (for the DB) and Pydantic (for data validation).

psycopg2-binary
# The "database driver" or "translator" that allows Python
# to communicate with a PostgreSQL database.


# --- Configuration & Utilities ---
python-dotenv
# A helper library to load your database credentials (DB_USER, DB_PASSWORD)
# from the .env file into your application's environment.

passlib[bcrypt]

openai
google-generativeai